// packages/figma-parser/src/full-figma-types.d.ts
export declare type ScaleMode = 'FILL' | 'FIT' | 'TILE' | 'STRETCH';
export declare type PaintTypeSolid = 'SOLID';
export declare type PaintTypeGraident = 'GRADIENT_LINEAR' | 'GRADIENT_RADIAL' | 'GRADIENT_ANGULAR' | 'GRADIENT_DIAMOND';
export declare type PaintTypeImage = 'IMAGE' | 'EMOJI';
export declare type TextType = 'TEXT';
export declare type PaintType = PaintTypeSolid | PaintTypeGraident | PaintTypeImage;
/**
 * how the layer blends with layers below
 */
export declare type BlendMode = 'PASS_THROUGH' /** (Only applicable to objects with children) */ | 'NORMAL'
  /** Darken: */
  | 'DARKEN' | 'MULTIPLY' | 'LINEAR_BURN' | 'COLOR_BURN'
  /** Lighten: */
  | 'LIGHTEN' | 'SCREEN' | 'LINEAR_DODGE' | 'COLOR_DODGE'
  /** Contrast: */
  | 'OVERLAY' | 'SOFT_LIGHT' | 'HARD_LIGHT'
  /** Inversion: */
  | 'DIFFERENCE' | 'EXCLUSION'
  /** Component: */
  | 'HUE' | 'SATURATION' | 'COLOR' | 'LUMINOSITY';
// ...
export interface Color {
  /** Red channel value, between 0 and 1 */
  readonly r: number;
  /** Green channel value, between 0 and 1 */
  readonly g: number;
  /** Blue channel value, between 0 and 1 */
  readonly b: number;
  /** Alpha channel value, between 0 and 1 */
  readonly a: number;
}
// ...
export interface Effect {
  /** Type of effect as a string enum */
  readonly type: 'INNER_SHADOW' | 'DROP_SHADOW' | 'LAYER_BLUR' | 'BACKGROUND_BLUR';
  /** Is the effect active? */
  readonly visible: boolean;
  /** Radius of the blur effect (applies to shadows as well) */
  readonly radius: number;
  readonly color?: Color;
  readonly blendMode?: BlendMode;
  readonly offset?: Vector2;
  readonly spread?: number;
}
/** A solid color, gradient, or image texture that can be applied as fills or strokes */
export interface Paint {
  /** Type of paint as a string enum */
  readonly type: PaintType;
  /**
   * Is the paint enabled?
   * @default true
   */
  readonly visible?: boolean;
  /**
   * Overall opacity of paint (colors within the paint can also have opacity
   * values which would blend with this)
   * @default 1
   */
  readonly opacity?: number;
  /** Solid color of the paint */
  readonly color?: Color;
  /**
   * How this node blends with nodes behind it in the scene
   * (see blend mode section for more details)
   */
  readonly blendMode: BlendMode;
  /**
   * This field contains three vectors, each of which are a position in
   * normalized object space (normalized object space is if the top left
   * corner of the bounding box of the object is (0, 0) and the bottom
   * right is (1,1)). The first position corresponds to the start of the
   * gradient (value 0 for the purposes of calculating gradient stops),
   * the second position is the end of the gradient (value 1), and the
   * third handle position determines the width of the gradient (only
   * relevant for non-linear gradients).
   *
   */
  readonly gradientHandlePositions?: ReadonlyArray<Vector2>;
  /**
   * Positions of key points along the gradient axis with the colors
   * anchored there. Colors along the gradient are interpolated smoothly
   * between neighboring gradient stops.
   */
  readonly gradientStops?: ReadonlyArray<ColorStop>;
  /** Image scaling mode */
  readonly scaleMode?: ScaleMode;
  /**
   * Affine transform applied to the image, only present if scaleMode is `STRETCH`
   */
  readonly imageTransform?: Transform;
  /**
   * Amount image is scaled by in tiling, only present if scaleMode is `TILE`
   */
  readonly scalingFactor?: number;
  /**
   * A reference to an image embedded in the file. To download the image using this reference,
   * use the GET file images endpoint to retrieve the mapping from image references to image URLs
   */
  readonly imageRef?: string;
  /**
   * A reference to the GIF embedded in this node, if the image is a GIF.
   * To download the image using this reference,
   * use the GET file images endpoint to retrieve the mapping from image references to image URLs
   */
  readonly gifRef?: string;
}
export interface Path {
  /** A sequence of path commands in SVG notation */
  readonly path: string;
  /** Winding rule for the path */
  readonly windingRule: 'EVENODD' | 'NONZERO';
}
export declare type Transform = ReadonlyArray<ReadonlyArray<number>>;
/** A 2d vector */
export interface Vector2 {
  /** X coordinate of the vector */
  readonly x: number;
  /** Y coordinate of the vector */
  readonly y: number;
}
/** A position color pair representing a gradient stop */
export interface ColorStop {
  /** Value between 0 and 1 representing position along gradient axis */
  readonly position: number;
  /** Color attached to corresponding position */
  readonly color: Color;
}
/** Metadata for character formatting */
export interface TypeStyle {
  /** Font family of text (standard name) */
  readonly fontFamily: string;
  /** PostScript font name */
  readonly fontPostScriptName: string;
  /** Space between paragraphs in px, 0 if not present */
  readonly paragraphSpacing?: number;
  /** Paragraph indentation in px, 0 if not present */
  readonly paragraphIndent?: number;
  /** Is text italicized? */
  readonly italic?: boolean;
  /** Numeric font weight */
  readonly fontWeight: 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
  /** Font size in px */
  readonly fontSize: number;
  /** Horizontal text alignment as string enum */
  readonly textAlignHorizontal: 'LEFT' | 'RIGHT' | 'CENTER' | 'JUSTIFIED';
  /** Vertical text alignment as string enum */
  readonly textAlignVertical: 'TOP' | 'CENTER' | 'BOTTOM';
  /** Space between characters in px */
  readonly letterSpacing: number;
  /** Paints applied to characters */
  readonly fills?: ReadonlyArray<Paint>;
  /** Line height in px */
  readonly lineHeightPx: number;
  /** Line height as a percentage of normal line height */
  readonly lineHeightPercent: number;
  /** The unit of the line height value specified by the user. */
  readonly lineHeightUnit: 'PIXELS' | 'FONT_SIZE_%' | 'INTRINSIC_%';
  /** Text casing applied to the node, default is the original casing */
  readonly textCase?: 'UPPER' | 'LOWER' | 'TITLE';
  /** Text decoration applied to the node, default is none */
  readonly textDecoration?: 'STRIKETHROUGH' | 'UNDERLINE';
  /** Line height as a percentage of the font size. Only returned when lineHeightPercent is not 100. */
  readonly lineHeightPercentFontSize?: number;
}